requirements.txt
Flask==0.10.1
Flask-SocketIO==0.4.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.5
gunicorn==19.1.1
itsdangerous==0.24
simplejson==3.6.5
wsgiref==0.1.2
